PPO: restructure and name changes
