+++ data/valid_files.json
@@ -1,3 +1,8 @@
 {
-    "example_scenario.json": []
+    "example_scenario.json": [],
+    "intersection_scenario.json": [
+        0,
+        34,
+        39
+    ]
 }
